refactor: `view`, `domain` 으로 이름 변경
