+++ __tests__/Car.test.js
@@ -1,4 +1,4 @@
-import Car from '../src/models/Car';
+import Car from '../src/domain/Car';
 
 describe('Car', () => {
   test('자동차는 이름을 가질 수 있고 전진할 수 있어야 한다.', () => {
